Refactor RC2014 Mini into gsz80.cpp
