+++ src/mame/mess.flt
@@ -875,7 +875,6 @@ rambo.cpp
 ravens.cpp
 rc702.cpp
 rc759.cpp
-rc2014mini.cpp
 rd100.cpp
 regence.cpp
 replicator.cpp
